requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
openai==1.10.0
requests==2.31.0
pillow==10.2.0
numpy==1.24.3
matplotlib==3.7.1
python-dotenv==1.0.0
ipywidgets==8.1.1
ipython==8.18.1
jupyter==1.0.0